requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
bidict==0.21.2
bleach==3.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
docutils==0.16
dop-python-pip-package==0.0.4
Flask==1.1.2
Flask-Cors==3.0.9
flask-mysql-connector==1.0.3
Flask-SocketIO==5.0.1
grpcio==1.34.0
grpcio-tools==1.34.0
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
keyring==21.7.0
MarkupSafe==1.1.1
more-itertools==8.6.0
mysql-connector-python==8.0.22
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pkginfo==1.6.1
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
Pygments==2.7.3
pyparsing==2.4.7
pytest==3.7.2
python-dateutil==2.8.1
python-engineio==4.1.0
python-socketio==5.1.0
pytz==2020.4
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.55.0
twine==3.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.0
